requirements.txt
flask==2.0.1
flask-cors==4.0.0
firebase-admin==6.0.0
google-generativeai==0.8.2
pydub==0.25.1
python-dotenv==0.19.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.0
google-cloud-logging==3.10.0
numpy==1.26.4
wavio==0.0.8
schedule==1.1.0
gunicorn==20.1.0
werkzeug==2.0.3
